ATMegaKeyScanner: Use _KeyScannerDescription::KeyAddr instead of global
To be able to define the key scanner at the same time we define its description,
we can't rely on the global `KeyAddr` anymore, we need to pick the one from the
description.

Since we want to be able to define both at the same time, lets do just that, and
pick the `KeyAddr` type from the key scanner description.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>