testing: Make cmake-clean not depend on clean
While we do want to run `make clean` if we _can_, it must not be a dependency,
but an optional part only. It must not be, because if we have it built on the
host, we can't run `make clean` in docker and vice versa - yet, we will want to
remove the cmake artifacts to be able to regenerate.

Technically, we don't even _need_ `make clean`, it just feels cleaner to have it.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>